Fix issue saving the session
Got me stuck for a while, but found it.
The issue is that sqllite dont autogenerate the UUIDs, so the id for the
user was always set to null (also I wonder, if it's primary key, why
accept the null). Anyway, I fixed that and a few other stuff in this
commit, such as showing the jwt properly in the web and son on.