Change sqllite schema UUIDs -> TEXT
Well it turns out that UUIDs are not fully supported in SQLlite, what a
bummer ... So I had to just get rid of them.
I still use the model to use UUIDs tho. If I switch to postgres at some
point, I could turn back to UUID that autogenerate, and update the
models again.